Extract an object responsible for creating machines
ControlPlaneUpgrader was doing a lot. One thing it was doing
was creating machines and waiting for the backing node to get to a certain
state.

In order to properly test idempotency we need to be able to stop the Upgrade
at some point. This refactor starts to allow us to have full control
over what actions the Upgrader is performing.

Format fixes and custom timeouts

Adds a test with aws-like provider ids

Signed-off-by: Chuck Ha <chuckh@vmware.com>