Edit kubeadm ConfigMap for control plane upgrades
When a new control plane node joins the cluster, it gets the version to
use as the image tag for the static pod manifests from the
kubeadm-config ConfigMap in the kube-system namespace. We must change
the ClusterConfiguration.kubernetesVersion to match the desired version,
then upload the updated ConfigMap before proceeding to create any new
machines.

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>