Extract ClusterConfiguration for JSON patching
Extract the ClusterConfiguration payload from the Kubeadm configmap in
the workload cluster so that we can apply JSON patches during upgrades.
This is required because the ClusterConfiguration object is stored as
Yaml inside the configmap. Once we convert the Yaml to the correct
RuntimeObject we can then apply our patches.

(cherry picked from commit 5f7e04357b11900aed298f89451525ca33900999)