Remove .upgrade. from machine names
Forcing replacment machine names to include ".upgrade." was breaking
some Cluster API providers that map the machine name directly to the
instance's hostname. This change removes that hardcoding and replaces it
with a single dash followed by whatever the user specifies for the
upgrade ID. Additionally, it keeps track of the "base name" of a
machine in an annotation, and uses that to generate the replacement
machine name instead of trying to search for a special suffix. This
means, for example, that a machine named "cp-0" can have an upgrade ID
of "0": the new name is cp-0-0, and the upgrade tool correctly handles
this now.

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>