Fall back to 127.0.0.1 in etcdctl command when pod has an unparseable IP
In certain situations, the PodIP may not only be set to the empty
string, but it can be set to "<nil>". It's unclear _why_ that happens,
but until there are validations that prevent these malformed IPs from
being set, the command needs to account for this.

Rather than trying to enumerate all of the likely malformed strings,
falling back in the event of a failed parse should cover any future
scenarios as well.